backend/requirements.txt
strawberry-graphql[fastapi]==0.209.2
uvicorn[standard]==0.23.2
sqlalchemy_utils==0.41.1
python-multipart==0.0.6
sse-starlette==1.6.5
sqlalchemy==2.0.21
starlette==0.27.0
pydantic==2.4.2
aiofiles==23.2.1
fastapi==0.103.2
alembic==1.12.0
pyirsdk==1.3.3
pylint==2.17.6
Pillow==10.0.1
black==23.9.1
colour==0.1.5
redis==5.0.1
ujson==5.8.0
sacn==1.9.0
